requirements.txt
opencv-python
insightface
onnxruntime
faiss-cpu
numpy
flask
flask-cors